// angular/src/app/interfaces/fotopersonal.ts
export interface fotopersonal {
    idper?: number;
    nombre: string;
    apellidos: string;
    usuario: string;
    contra: string;
    fechanac: string;
    sexo: string;
    curp: string;
    estadocivil: string;
    domicilio: string;
    fechaingreso: string;
    horasalida: string;
    horaentrada: string;
    perfilprofesional: string;
    especialidad: string;
    salariomensual: number;
    tareasasignadas: string;
    idtper: number;
    activo: number;
    fotopersonal:number;
}